rewrite debug module
