--- src/main/kotlin/io/xmake/debug/DefaultDebugConfigurations.kt
@@ -1,226 +1,0 @@
-package io.xmake.debug
-
-import com.intellij.openapi.project.Project
-import com.intellij.execution.configurations.GeneralCommandLine
-import kotlinx.serialization.Serializable
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.JsonElement
-import kotlinx.serialization.json.JsonObject
-import kotlinx.serialization.json.jsonObject
-import kotlinx.serialization.json.jsonPrimitive
-import kotlinx.serialization.json.decodeFromJsonElement
-import kotlinx.serialization.json.JsonPrimitive
-import kotlinx.serialization.json.JsonArray
-import io.xmake.utils.Logger
-
-/**
- * Default debug configurations for different DAP drivers
- */
-object DefaultDebugConfigurations {
-    
-    private const val TAG = "DefaultDebugConfigurations"
-    
-    // Default DAP configuration
-    val defaultDapConfig = mapOf(
-        "stopOnEntry" to true,
-        "sourceMap" to mapOf(
-            "enabled" to "true"
-        ),
-        "showDisassembly" to "auto",
-        "enablePrettyPrinting" to true,
-        "timeout" to 30000,
-        "memoryReference" to "hex",
-        "displayFormat" to "hex",
-        "maxChildren" to 1000,
-        "maxArrayLength" to 1000,
-        "maxStringLength" to 10000,
-        "initCommands" to emptyList<String>(),
-        // Variable display settings
-        "variables" to mapOf(
-            "showArguments" to true,
-            "showLocals" to true,
-            "showGlobals" to true,
-            "showStatics" to true,
-            "showRegisters" to true
-        )
-    )
-    
-    // LLDB-specific default configuration
-    val defaultLldbConfig = mapOf(
-        "initCommands" to listOf(
-            "settings set target.inline-breakpoint-strategy always",
-            "settings set target.process.memory-protection-rules false",
-            "settings set target.process.memory-cache-line-size 64",
-            "settings set target.process.disable-aslr false",
-            "settings set target.process.detach-on-error false",
-            "settings set target.process.stop-on-sharedlib-events false",
-            "settings set target.process.stop-on-sharedlib-load-state false",
-            "settings set target.process.stop-on-plugin-load false",
-            "settings set target.process.stop-on-dylib-load false",
-            "settings set target.process.stop-on-dylib-unload false",
-            "settings set target.process.stop-on-objc-throw false",
-            "settings set target.process.stop-on-cxx-exception-break false",
-            "settings set target.process.stop-on-cxx-exception-throw false",
-            "settings set target.process.stop-on-catch-throw false",
-            "settings set target.process.stop-on-objc-exception-bp false",
-            "settings set target.process.stop-on-cxx-exception-bp false",
-            "settings set target.process.stop-on-catch-bp false",
-            "settings set target.process.stop-on-exception-bp false",
-            "settings set target.process.stop-on-exception-throw false",
-            "settings set target.process.stop-on-exception-catch false",
-            "settings set target.process.stop-on-exception-rethrow false",
-            "settings set target.process.stop-on-exception-sigsegv false",
-            "settings set target.process.stop-on-exception-sigbus false",
-            "settings set target.process.stop-on-exception-sigfpe false",
-            "settings set target.process.stop-on-exception-sigill false",
-            "settings set target.process.stop-on-exception-sigtrap false",
-            "settings set target.process.stop-on-exception-sigint false",
-            "settings set target.process.stop-on-exception-sigterm false",
-            "settings set target.process.stop-on-exception-sigkill false",
-            // Memory and variable display settings
-            "settings set target.memory-module-load-swift-info yes",
-            "settings set target.memory-module-load-rust-info yes",
-            "settings set target.memory-module-load-go-info yes",
-            "settings set target.memory-module-load-dlang-info yes",
-            "settings set target.memory-module-load-nim-info yes",
-            "settings set target.memory-module-load-vb-info yes",
-            "settings set target.memory-module-load-fsharp-info yes",
-            "settings set target.memory-module-load-powershell-info yes",
-            "settings set target.memory-module-load-python-info yes",
-            "settings set target.memory-module-load-ruby-info yes",
-            "settings set target.memory-module-load-perl-info yes",
-            "settings set target.memory-module-load-php-info yes",
-            "settings set target.memory-module-load-javascript-info yes",
-            "settings set target.memory-module-load-typescript-info yes",
-            "settings set target.memory-module-load-coffeescript-info yes",
-            "settings set target.memory-module-load-bash-info yes",
-            "settings set target.memory-module-load-zsh-info yes",
-            "settings set target.memory-module-load-fish-info yes",
-            "settings set target.memory-module-load-csh-info yes",
-            "settings set target.memory-module-load-tcsh-info yes",
-            "settings set target.memory-module-load-ksh-info yes",
-            "settings set target.memory-module-load-sh-info yes"
-        )
-    )
-    
-    // GDB-specific default configuration
-    val defaultGdbConfig = mapOf(
-        "initCommands" to listOf(
-            "set pagination off",
-            "set print pretty on",
-            "set print array on",
-            "set print elements 0",
-            "set print null-stop on",
-            "set confirm off",
-            "set unwindonsignal on",
-            "set step-mode on",
-            "set print object on",
-            "set print static-members on",
-            "set print vtbl on",
-            "set print demangle on",
-            "set demangle-style gnu-v3",
-            "set print sevenbit-strings off",
-            "set print union on",
-            "set print repeats 0",
-            "set print frame-arguments all",
-            "set print frame-info auto",
-            "set print entry-values no",
-            "set print raw-frame-arguments no"
-        )
-    )
-    
-    /**
-     * Get default configuration for specific driver type
-     */
-    fun getDefaultConfig(driverType: DapDriverDetector.DapDriverType): Map<String, Any> {
-        return when (driverType) {
-            DapDriverDetector.DapDriverType.LLDB_DAP -> 
-                defaultDapConfig + defaultLldbConfig
-            DapDriverDetector.DapDriverType.GDB_DAP -> 
-                defaultDapConfig + defaultGdbConfig
-            else -> defaultDapConfig
-        }
-    }
-    
-    /**
-     * Parse user configuration and merge with defaults using safe JSON parsing
-     */
-    fun mergeConfigurations(
-        userConfigJson: String,
-        defaultConfig: Map<String, Any>
-    ): Map<String, Any> {
-        if (userConfigJson.isBlank()) {
-            return defaultConfig
-        }
-        
-        return try {
-            val json = Json { ignoreUnknownKeys = true }
-            val userConfig = json.decodeFromString<JsonObject>(userConfigJson)
-            
-            val mergedConfig = defaultConfig.toMutableMap()
-            
-            userConfig.forEach { (key, value) ->
-                when (value) {
-                    is JsonObject -> {
-                        // Handle nested objects like sourceMap
-                        val existingNested = (mergedConfig[key] as? Map<String, Any>) ?: emptyMap<String, Any>()
-                        val userNested = value.mapValues { 
-                            convertJsonElement(it.value, key == "sourceMap" && it.key == "enabled")
-                        }
-                        mergedConfig[key] = existingNested + userNested
-                    }
-                    else -> {
-                        // Handle primitive values using safe conversion
-                        mergedConfig[key] = convertJsonElement(value)
-                    }
-                }
-            }
-            
-            mergedConfig
-        } catch (e: Exception) {
-            Logger.e(TAG, "Failed to parse user launch configuration JSON: ${e.message}", e)
-            Logger.d(TAG, "Invalid JSON content: $userConfigJson")
-            // If JSON parsing fails, return default config
-            defaultConfig
-        }
-    }
-    
-    /**
-     * Safely convert JsonElement to appropriate type
-     */
-    private fun convertJsonElement(element: JsonElement, forceString: Boolean = false): Any {
-        return when {
-            forceString -> {
-                // Special case: sourceMap.enabled should remain as string
-                when (element) {
-                    is JsonPrimitive -> element.content
-                    else -> element.toString()
-                }
-            }
-            element is JsonPrimitive -> {
-                when {
-                    element.content == "true" || element.content == "false" -> element.content.toBoolean()
-                    element.content.toLongOrNull() != null -> {
-                        // Try to preserve integer type when possible
-                        val longValue = element.content.toLong()
-                        if (longValue.toInt().toLong() == longValue) {
-                            longValue.toInt()
-                        } else {
-                            longValue
-                        }
-                    }
-                    else -> element.content
-                }
-            }
-            element is JsonArray -> {
-                // Convert JSON array to List<Any>
-                element.map { convertJsonElement(it) }
-            }
-            element is JsonObject -> {
-                // Convert nested object to Map<String, Any>
-                element.mapValues { convertJsonElement(it.value) }
-            }
-            else -> element.toString()
-        }
-    }
-}
