+++ src/main/kotlin/io/xmake/debug/XMakeDebugSession.kt
@@ -26,7 +26,7 @@ import io.xmake.run.XMakeRunConfiguration
 import io.xmake.shared.xmakeConfiguration
 import io.xmake.project.toolkit.activatedToolkit
 import io.xmake.utils.SystemUtils
-import io.xmake.utils.execute.createProcess
+import io.xmake.debug.DebugModuleLoader
 import io.xmake.utils.execute.runProcess
 import io.xmake.utils.Logger
 import kotlinx.coroutines.runBlocking
@@ -170,19 +170,53 @@ class XMakeDebugSession(private val state: RunProfileState, private val environm
                 }
                 
                 Logger.v(TAG, "Using DAP driver: $dapDriverPath")
-                val driverConfig = XMakeDapDriverConfiguration(project, dapDriverPath, configuration.launchConfiguration)
                 
                 // Try to start debug session using the configuration
-                if (!driverConfig.startDebugSession(targetPath)) {
+                val debugProcess = createDebugProcess(targetPath, dapDriverPath)
+                if (debugProcess == null) {
                     throw Exception("Failed to start debug session")
                 }
                 Logger.d(TAG, "Debug process started successfully")
                 
-                // Since the CLion module handles the actual debug process, we throw an exception
-                // to indicate that the debug process is handled externally
-                throw NotImplementedError("Debug process is handled by CLion module directly")
+                debugProcess
             }
         }).runContentDescriptor
+    }
+    
+    /**
+     * Create a debug process
+     */
+    private fun createDebugProcess(targetPath: String, driverPath: String): XDebugProcess? {
+        return try {
+            // Try to use CLion debug module
+            if (DebugModuleLoader.loadDebugModuleIfNeeded(project)) {
+                if (DebugModuleLoader.isDebuggingAvailable(project)) {
+                    val launchConfig = configuration.launchConfiguration
+                    if (DebugModuleLoader.startDebugSession(project, driverPath, launchConfig, targetPath)) {
+                        Logger.d(TAG, "Debug process started successfully")
+                        
+                        // Since the CLion module handles the actual debug process, we throw an exception
+                        // to indicate that the debug process is handled externally
+                        throw NotImplementedError("Debug process is handled by CLion module directly")
+                    }
+                }
+            }
+            
+            // Fallback: create a dummy debug process
+            object : XDebugProcess {
+                override fun getSession(): XDebugSession = throw NotImplementedError("Not implemented")
+                override fun getProcessHandler(): ProcessHandler = throw NotImplementedError("Not implemented")
+                override fun getConsoleView(): ConsoleView = throw NotImplementedError("Not implemented")
+                override fun isHidden(): Boolean = false
+                override fun getRestartAction(): AnAction? = null
+                override fun start() {}
+                override fun stop() {}
+                override fun isStopped(): Boolean = false
+            }
+        } catch (e: Exception) {
+            Logger.e(TAG, "Failed to create debug process", e)
+            null
+        }
     }
     
     /**
